Add decorator with campaigns without alter decorated class Add Discount campaign Add dynamic campaign and make easy to create more Campaigns Add Integrated and Unit test for new campaign

+++ src/main/java/io/jcervelin/fashionstore/cart/domains/CampaignAttributes.java
@@ -21,7 +21,7 @@ public class CampaignAttributes {
     private String campaignType;
     private LocalDateTime starts;
     private LocalDateTime expires;
-    private Double percentFactor;
+    private Integer percentFactor;
     private Integer buyX;
     private Integer toY;
     private List<Product> productsAffected;
